+++ backend/src/main/java/com/fitsync/config/SecurityConfig.java
@@ -12,6 +12,8 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
@@ -22,7 +24,7 @@ import java.util.List;
 
 @Configuration // 스프링의 설정 클래스임을 나타냅니다
 @EnableWebSecurity // Spring Security 활성화
-@RequiredArgsConstructor // final 필드에 대한 생성자를 자동으로 생성
+@RequiredArgsConstructor
 public class SecurityConfig {
 
     // OAuth2 인증 처리를 위한 커스텀 서비스
@@ -39,6 +41,12 @@ public class SecurityConfig {
             "/", "/css/**", "/images/**", "/js/**", "/favicon.ico", "/error",
             "/oauth2/**", "/api/auth/**", "/health" // OAuth2 관련 경로와 토큰 재발급 등의 인증 관련 경로
     };
+
+    // 비밀번호 단방향(BCrypt) 암호화용 Bean
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
 
     @Bean
     public CorsConfigurationSource corsConfigurationSource() {
